Fix website role (fixes #25, #26, #16)
This commit is almost a complete rewrite of the website role, that uses
Caddy to host the website, checkouts and builds the website once new
contect (on master branch) is detected (on a 5 minute interval rate).
Onion services version 2 and 3 are now working and use a different
network than localhost for accidental leaks or issues with localhost
bidning of onion services.

Resolves:
* https://github.com/PrivacyLx/privacylx-issue-tracker/issues/16
* https://github.com/PrivacyLx/privacylx-issue-tracker/issues/25
* https://github.com/PrivacyLx/privacylx-issue-tracker/issues/26

* No need for submodules, update README
* Rename caddy-website to website
* Fix order of deployment roles, docker should be before website
* Move some known public variables to group_vars
* Cleanup: remove broken, obsolete and other non required files
* Update privacylx.org host variables, SSH known_hosts, and inventory