Testing of Discourse role with Travis and Vagrant
Support for testing Ansible roles with Molecule. The Vagrant platform is
being used in order to be able and test Ansible playbook roles that
deploy and use Docker containers. Molecule does not support Docker
inside Docker testing and Vagrant help us to test our roles as accurate
as possible without workarounds or extra modifications.

This commit introduces default variables required by Molecule which is
useful for others in case they want to use the Discourse role in their
setup. Additionally it's fixing the faulty conditionals in Discourse
role. Furthermore it adds an Asnible inventory specifically for Molecule
testing.

Molecule does linting and syntax checking, deploys the playbook role,
checks if playbook is idempotent and finally runs some preliminary
tests; checks if Docker daemon is up and running on the testing
platforms.